+++ src/main/java/com/dh/sistemadereservadeturnos/service/implementation/PacienteService.java
@@ -1,8 +1,10 @@
 package com.dh.sistemadereservadeturnos.service.implementation;
 
 import com.dh.sistemadereservadeturnos.entity.Paciente;
+import com.dh.sistemadereservadeturnos.exception.ResourceNotFoundException;
 import com.dh.sistemadereservadeturnos.repository.IPacienteRepository;
 import com.dh.sistemadereservadeturnos.service.IPacienteService;
+import org.apache.log4j.Logger;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,6 +13,8 @@ import java.util.Optional;
 
 @Service
 public class PacienteService implements IPacienteService {
+
+    private static Logger logger = Logger.getLogger(PacienteService.class);
 
     private IPacienteRepository pacienteRepository;
 
@@ -32,19 +36,34 @@ public class PacienteService implements IPacienteService {
     @Override
     public Paciente buscarPorId(Long id) {
 
-        Optional<Paciente> pacienteOptional = pacienteRepository.findById(id);
-        return pacienteOptional.orElse(null);
+        Optional<Paciente> pacienteBuscado = pacienteRepository.findById(id);
+        if (pacienteBuscado.isEmpty()){
+            logger.error("Paciente no encontrado con id: " +id);
+            throw new ResourceNotFoundException("Paciente no encontrado con id: " +id);
+        }
+        return pacienteBuscado.get();
 
     }
 
     @Override
     public void eliminar(Long id) {
+
+        Optional<Paciente> pacienteAEliminar = pacienteRepository.findById(id);
+        if (pacienteAEliminar.isEmpty()){
+            logger.error("Paciente a eliminar no encontrado con id: " +id);
+            throw new ResourceNotFoundException("Paciente a eliminar no encontrado con id: " +id);
+        }
         pacienteRepository.deleteById(id);
     }
 
     @Override
     public void actualizar(Paciente paciente) {
-       pacienteRepository.save(paciente);
+        Optional<Paciente> pacienteAActualizar = pacienteRepository.findById(paciente.getId());
+     if (pacienteAActualizar.isEmpty()){
+         logger.error("Paciente a actualizar no encontrado con id: " +paciente.getId());
+         throw new ResourceNotFoundException("Paciente a actualizar no encontrado con id: " +paciente.getId());
+     }
+     pacienteRepository.save(paciente);
 
 
     }
